+++ gen/python/platform_contracts/app_studio/__init__.py
@@ -1,0 +1,3 @@
+from . import app_mission, app_task_plan
+
+__all__ = ["app_mission", "app_task_plan"]